+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `ngl_node_param_swap_elem` to swap elements within a list
 - `Group.children` can be live re-ordered
 - `ngl_update()` to trigger graph updates without producing frames
+- `ngl_jni_get_env()` to attach and get a JNI environment from any thread
 
 ### Fixed
 - Crash when using resizable RTTs with time ranges